multimodal-bookstore/requirements.txt
Flask==2.3.2
Pillow
pytesseract
requests
python-dotenv
openai
faiss-cpu
sentence-transformers